second release, fix negative number validation and change ID configuration

+++ WildlifeTracker/Mammals/Donkeys/Donkey.cs
@@ -9,14 +9,14 @@ namespace WildlifeTracker.Mammals.Donkeys
     public class Donkey : Mammal
     {
         #region // Instance variables //
-        private int maxLoad;
+        private double maxLoad;
         private bool isUsedAsPackAnimal;
         private double height;
         private double weight;
         #endregion
 
         #region // Properties //
-        public int MaxLoad { get => maxLoad; set => maxLoad = value; }
+        public double MaxLoad { get => maxLoad; set => maxLoad = value; }
         public bool IsUsedAsPackAnimal { get => isUsedAsPackAnimal; set => isUsedAsPackAnimal = value; }
         public double Height { get => height; set => height = value; }
         public double Weight { get => weight; set => weight = value; }
